backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==6.0.1
requests==2.31.0
python-dotenv==1.0.0
Pillow>=10.0.0
flask_socketio
eventlet